template_requirements.txt
discord.py
certifi
pip-system-certs
ruamel.yaml
PyNaCl
python-box
pydantic
tortoise-orm
aiosqlite